Move prepare data related modules to preparation
