+++ tests/test_modules/test_featurecreator.py
@@ -2,7 +2,7 @@ import unittest
 import pandas as pd
 import dask.dataframe as dd
 from datetime import datetime
-from corebehrt.modules.features import FeatureCreator
+from corebehrt.modules.preprocessing.features import FeatureCreator
 
 
 class TestFeatureCreator(unittest.TestCase):
